Multiple ball count
Vector of <ball>. Each time contours are detected, they are mapped to nearest
balls. Those balls' centers are then updated with the most recent contour
center.

If there are more contours than balls, then new balls are initialized and
mapped for the extra contours.

If there are more balls than contours, only the closest balls are kept. The
other balls are assumed to have been caught, and are removed from the vector.